som innan  fast med en login rutan som huvudpjäs
