--- src/main/java/org/example/TimerUI.java
@@ -1,4 +1,0 @@
-package org.example;
-
-public class TimerUI {
-}
